requirements.txt
# Python 3.8+
requests>=2.31.0
python-dotenv>=1.0.0
